requirements.txt
matplotlib
seaborn
pandas
python-docx
docx2pdf